// src/client/src/constants/colors.ts
export enum Color {
    WHITE = 'white',
    BLACK100 = '#F5F5F6',
    BLACK200 = '#D5D6DA',
    BLACK300 = '#82858F',
    BLACK400 = '#595D69',
    BLACK500 = '#2F3444',
    BLUE100 = '#7FA0FF',
    BLUE200 = '#5580FF',
    BLUE300 = '#2A60FF',
    BLUE400 = '#224DCC',
    BLUE500 = '#193A99',
    ORANGE100 = '#FAA783',
    ORANGE200 = '#F98959',
    ORANGE300 = '#F76C30',
    ORANGE400 = '#C65626',
    ORANGE500 = '#94411D',
    RED100 = '#FF8383',
    RED200 = '#FF5A5A',
    RED300 = '#FF2B2B',
    RED400 = '#CC2727',
    RED500 = '#991D1D',
    GREEN100 = '#66E983',
    GREEN200 = '#33E159',
    GREEN300 = '#00DA30',
    GREEN400 = '#03AE26',
    GREEN500 = '#06841E',
}